+++ src/main/java/common/legobmw99/allomancy/items/ItemCoinBag.java
@@ -4,8 +4,10 @@ import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.init.Items;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
+import net.minecraft.util.ActionResult;
+import net.minecraft.util.EnumActionResult;
+import net.minecraft.util.EnumHand;
 import net.minecraft.world.World;
-
 import common.legobmw99.allomancy.common.AllomancyCapabilities;
 import common.legobmw99.allomancy.common.Registry;
 import common.legobmw99.allomancy.entity.EntityGoldNugget;
@@ -19,17 +21,18 @@ public class ItemCoinBag extends Item{
 		}
 	
 
-	public ItemStack onItemRightClick(ItemStack par1ItemStack, World par2World, EntityPlayer par3EntityPlayer)
-    {
-        ItemStack itemstack = this.findArrow(par3EntityPlayer);
-        if (par3EntityPlayer.capabilities.isCreativeMode || itemstack != null && AllomancyCapabilities.forPlayer(par3EntityPlayer).MetalBurning[AllomancyCapabilities.matSteel]){
-        		EntityGoldNugget entitygold = new EntityGoldNugget(par2World, par3EntityPlayer);
-        		par2World.spawnEntityInWorld(entitygold);
-        		if(!par3EntityPlayer.capabilities.isCreativeMode){
+	   public ActionResult<ItemStack> onItemRightClick(ItemStack itemStackIn, World worldIn, EntityPlayer playerIn, EnumHand hand)
+	    {
+	  
+        ItemStack itemstack = this.findArrow(playerIn);
+        if (playerIn.capabilities.isCreativeMode || itemstack != null && AllomancyCapabilities.forPlayer(playerIn).MetalBurning[AllomancyCapabilities.matSteel]){
+        		EntityGoldNugget entitygold = new EntityGoldNugget(worldIn, playerIn);
+        		worldIn.spawnEntityInWorld(entitygold);
+        		if(!playerIn.capabilities.isCreativeMode){
         			--itemstack.stackSize;
         		}
         	}        
-        return par1ItemStack;
+        return new ActionResult(EnumActionResult.PASS, itemStackIn);
     }
 
 	/*
